Перенёс функцию фильтрации изображений в отдельный модуль
